templates: make template handling provider agnostic
This PR moves template handling code outside of oVirt provider.
