+++ pkg/templates/templates_suite_test.go
@@ -1,0 +1,13 @@
+package templates_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestTemplates(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Templates Suite")
+}
